requirements.txt
yapf==0.30.0
tensorflow==2.3.1
tensorflow-probability==0.11.1
tensorboardX
metadrive-simulator==0.2.3
loguru
imageio
easydict
tensorboardX
pyyaml
gym==0.18.0
ray[all]==1.0.0
stable_baselines3
protobuf==3.20.0